code refactoring as per the PR comments
