+++ src/main/resources/application.properties
@@ -8,6 +8,4 @@ barcode.url=${BARCODE_SERVICE_URL}
 
 api.url=${API_URL}
 
-spring.jackson.deserialization.fail-on-unknown-properties=true
-
 logging.file.flush=true
